+++ CHANGELOG.rst
@@ -1,7 +1,8 @@
 Unreleased
 -----
-* CLI: Improved 'Set' validator to allow non-string types, and CSV params ((`#452 <https://github.com/tomerfiliba/plumbum/pull/452>`_)
-* Commands: execute local/remote commands via a magic `.cmd` attribute ((`#450 <https://github.com/tomerfiliba/plumbum/pull/450>`_)
+* CLI: Improved 'Set' validator to allow non-string types, and CSV params (`#452 <https://github.com/tomerfiliba/plumbum/pull/452>`_)
+* TypedEnv: Facility for modeling environment-variables into python data types (`#451 <https://github.com/tomerfiliba/plumbum/pull/450>`_)
+* Commands: execute local/remote commands via a magic `.cmd` attribute (`#450 <https://github.com/tomerfiliba/plumbum/pull/450>`_)
 
 1.6.7
 -----